Check if hardware_info has been updated or not.
The hardware_info struct is allocated somewhere else and the update_hardware_info_8x16
method will set the structure field. In case of an unknown device,
(e.g. msm8939-florida-snd-card for Motorola lux) the structure is not
initialized. Returning that the structure was not updated allows caller
to release hardware_info struct and set it to NULL.

Current behavior leads to random memory accesses and crashes by hw_info_append_hw_type

Change-Id: I20c067c67de2b1e6a0294fa00d64a64484c7eb54